House-cleaning for files, added print statements to track flow of method calls, and changed flag variables for testing
