--- checkpoint
@@ -1,2 +1,0 @@
-model_checkpoint_path: "DCGAN.model-60502"
-all_model_checkpoint_paths: "DCGAN.model-60502"
