+++ dcgan/checkpoint/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "DCGAN.model-2"
+all_model_checkpoint_paths: "DCGAN.model-2"
